minor fixes

+++ src/perf.h
@@ -2,6 +2,7 @@
 #ifndef PERF_H
 #define PERF_H
 
+#include <cstdint>
 #include <iostream>
 #include <chrono>
 #include <utility>
@@ -10,8 +11,8 @@
 
 using perf_clock = std::chrono::_V2::high_resolution_clock::time_point;
 using perf_time = std::chrono::duration<double>;
-using perf_ms_time = std::chrono::duration<double, std::milli>;
-using perf_ns_time = std::chrono::duration<double, std::nano>;
+using perf_ms_time = std::chrono::duration<int64_t, std::milli>;
+using perf_ns_time = std::chrono::duration<int64_t, std::nano>;
 
 namespace perf
 {
